backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
SQLAlchemy
passlib[bcrypt]
python-jose
python-jose
passlib